+++ frontend/src/features/material-details/components/StockModal.tsx
@@ -19,23 +19,30 @@ under the License.
 SPDX-License-Identifier: Apache-2.0
 */
 
-import { useEffect, useState } from 'react';
-import { Input, Table } from '@catena-x/portal-shared-components';
-import { UNITS_OF_MEASUREMENT } from '@models/constants/uom';
+import { Table } from '@catena-x/portal-shared-components';
+import { Box, Button, capitalize, Dialog, DialogTitle, Stack } from '@mui/material';
+import { Close, Delete, Edit } from '@mui/icons-material';
 import { Stock, StockType } from '@models/types/data/stock';
-import { Box, Button, Checkbox, Dialog, DialogTitle, FormLabel, Grid, InputLabel, Stack, capitalize } from '@mui/material';
-import { getUnitOfMeasurement, isValidOrderReference } from '@util/helpers';
-import { usePartners } from '@features/stock-view/hooks/usePartners';
-import { Close, Delete, Save } from '@mui/icons-material';
-import { ModalMode } from '@models/types/data/modal-mode';
-import { LabelledAutoComplete } from '@components/ui/LabelledAutoComplete';
-import { GridItem } from '@components/ui/GridItem';
-import { useSites } from '@features/stock-view/hooks/useSites';
+import { deleteStocks } from '@services/stocks-service';
 import { useNotifications } from '@contexts/notificationContext';
-import { deleteStocks, postStocks } from '@services/stocks-service';
+import { useDataModal } from '@contexts/dataModalContext';
+import { getUnitOfMeasurement } from '@util/helpers';
+import { DirectionType } from '@models/types/erp/directionType';
 
-const createStockColumns = (handleDelete?: (row: Stock) => void) => {
-    const columns = [
+export type StockModalProps = {
+    open: boolean;
+    onClose: () => void;
+    onRemove?: (uuid: string) => void;
+    stock: Partial<Stock> | null;
+    stocks: Stock[];
+    stockType: StockType;
+    onSave: (d?: Stock) => void;
+};
+
+const createStockColumns = (
+    handleDelete: (row: Stock) => void, handleEdit: (row: Stock) => void
+) => {
+    return [
         {
             field: 'quantity',
             headerName: 'Quantity',
@@ -48,16 +55,16 @@ const createStockColumns = (handleDelete?: (row: Stock) => void) => {
             ),
         },
         {
-          field: 'stockLocationBpns',
-          headerName: 'Site',
-          headerAlign: 'center',
-          flex: 2,
-          renderCell: (data: { row: Stock }) => (
-              <Box display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
-                  {data.row.stockLocationBpns}
-              </Box>
-          ),
-      },
+            field: 'stockLocationBpns',
+            headerName: 'Site',
+            headerAlign: 'center',
+            flex: 2,
+            renderCell: (data: { row: Stock }) => (
+                <Box display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
+                    {data.row.stockLocationBpns}
+                </Box>
+            ),
+        },
         {
             field: 'partner',
             headerName: 'Partner',
@@ -110,286 +117,99 @@ const createStockColumns = (handleDelete?: (row: Stock) => void) => {
                 </Box>
             ),
         },
-    ] as const;
-    if (handleDelete) {
-        return [
-            ...columns,
-            {
-                field: 'delete',
-                headerName: '',
-                sortable: false,
-                disableColumnMenu: true,
-                headerAlign: 'center',
-                type: 'string',
-                width: 30,
-                renderCell: (data: { row: Stock }) => {
-                    return (
-                        <Box display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
-                            <Button variant="text" color="error" onClick={() => handleDelete(data.row)} data-testid="delete-stock">
-                                <Delete></Delete>
-                            </Button>
-                        </Box>
-                    );
-                },
+        {
+            field: 'edit',
+            headerName: '',
+            sortable: false,
+            disableColumnMenu: true,
+            width: 20,
+            renderCell: (data: { row: Stock }) => {
+                return (
+                    <Box display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
+                        <Button variant="text" onClick={() => handleEdit(data.row)} data-testid="edit-stock">
+                            <Edit />
+                        </Button>
+
+                    </Box>
+                );
             },
-        ] as const;
-    }
-    return columns;
+        },
+        {
+            field: 'delete',
+            headerName: '',
+            width: 40,
+            sortable: false,
+            disableColumnMenu: true,
+            renderCell: ({ row }: { row: Stock }) => (
+                <Box display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
+
+                    <Button
+                        variant="text"
+                        color="error"
+                        onClick={() => handleDelete(row)}
+                    >
+                        <Delete />
+                    </Button>
+                </Box>
+            ),
+        },
+    ] as const;
 };
 
-type StockModalProps = {
-    open: boolean;
-    mode: ModalMode;
-    onClose: () => void;
-    onSave: () => void;
-    onRemove?: (deletedUuid: string) => void;
-    stock: Partial<Stock> | null;
-    stocks: Stock[];
-    stockType: StockType;
-};
-const isValidStock = (stock: Partial<Stock>) =>
-    stock &&
-    typeof stock.quantity === 'number' && stock.quantity >= 0 &&
-    stock.measurementUnit &&
-    stock.partner &&
-    stock.stockLocationBpns &&
-    stock.stockLocationBpna &&
-    isValidOrderReference(stock);
-
-export const StockModal = ({ open, mode, onClose, onSave, onRemove, stock, stocks, stockType }: StockModalProps) => {
-    const [temporaryStock, setTemporaryStock] = useState<Partial<Stock>>(stock ?? {});
-    const { partners } = usePartners(stockType, temporaryStock?.material?.ownMaterialNumber ?? null);
-    const { sites } = useSites();
+export const StockModal = ({
+    open,
+    onClose,
+    onRemove,
+    stock,
+    stocks,
+    stockType,
+}: StockModalProps) => {
     const { notify } = useNotifications();
-    const [formError, setFormError] = useState(false);
+    const { openDialog } = useDataModal();
+    const direction = stockType == 'material' ? DirectionType.Inbound : DirectionType.Outbound
 
-    const handleSaveClick = () => {
-        temporaryStock.customerOrderNumber ||= undefined;
-        temporaryStock.customerOrderPositionNumber ||= undefined;
-        temporaryStock.supplierOrderNumber ||= undefined;
-        if (!isValidStock(temporaryStock)) {
-            setFormError(true);
-            return;
-        }
-        setFormError(false);
-        postStocks(stockType, temporaryStock)
-            .then(() => {
-                onSave();
-                notify({
-                    title: 'Stock Created',
-                    description: 'The Stock has been saved successfully',
-                    severity: 'success',
-                });
-            })
-            .catch((error) => {
-                notify({
-                    title: error.status === 409 ? 'Conflict' : 'Error requesting update',
-                    description: error.status === 409 ? 'Date conflicting with another Stock' : error.error,
-                    severity: 'error',
-                });
-            })
-            .finally(onClose);
+    const handleEdit = (row: Stock) => {
+        openDialog('stock', { ...row }, stocks, 'edit', direction);
     };
     const handleDelete = async (row: Stock) => {
         if (row.uuid) {
             try {
                 await deleteStocks(stockType, row.uuid);
                 onRemove?.(row.uuid);
-            } catch (error) {
                 notify({
-                    title: 'Error deleting stock',
-                    description: 'Failed to delete the stock',
+                    title: 'Stock deleted',
+                    description: 'The stock record was successfully deleted.',
+                    severity: 'success',
+                });
+            } catch (e) {
+                notify({
+                    title: 'Error',
+                    description: 'Failed to delete stock record.',
                     severity: 'error',
                 });
             }
         }
     };
-    useEffect(() => {
-        if (stock) setTemporaryStock(stock);
-    }, [stock]);
+
     return (
         <>
             <Dialog open={open && stock !== null} onClose={onClose} data-testid="stock-modal">
                 <DialogTitle variant="h3" textAlign="center">
-                    {capitalize(mode)} {capitalize(stockType)} Stock
+                    {capitalize('view')} {capitalize(stockType)} Stock
                 </DialogTitle>
                 <Stack padding="0 2rem 2rem" sx={{ width: '60rem' }}>
-                    <Grid container spacing={2} padding=".25rem">
-                        {mode === 'create' ? (
-                            <>
-                                <GridItem label="Material Number" value={temporaryStock.material?.ownMaterialNumber ?? ''} />
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        sx={{ margin: '0' }}
-                                        id="partner"
-                                        options={partners ?? []}
-                                        getOptionLabel={(option) => option?.name ?? ''}
-                                        label="Partner*"
-                                        placeholder="Select a Partner"
-                                        error={formError && !temporaryStock?.partner}
-                                        onChange={(_, value) => setTemporaryStock({ ...temporaryStock, partner: value ?? undefined })}
-                                        value={temporaryStock.partner ?? null}
-                                        isOptionEqualToValue={(option, value) => option?.uuid === value?.uuid}
-                                        data-testid="stock-partner-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        id="stockLocationBpns"
-                                        options={sites ?? []}
-                                        getOptionLabel={(option) => option.name ?? ''}
-                                        error={formError}
-                                        isOptionEqualToValue={(option, value) => option?.bpns === value.bpns}
-                                        onChange={(_, value) =>
-                                            setTemporaryStock({ 
-                                                ...temporaryStock,
-                                                stockLocationBpns: value?.bpns ?? undefined,
-                                                stockLocationBpna: value ? temporaryStock.stockLocationBpna : undefined 
-                                            })
-                                        }
-                                        value={sites?.find((s) => s.bpns === temporaryStock.stockLocationBpns) ?? null}
-                                        label="Stock Site*"
-                                        placeholder="Select a Site"
-                                        data-testid="stock-site-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        id="stockLocationBpna"
-                                        options={sites?.find(site => site.bpns === temporaryStock.stockLocationBpns)?.addresses ?? []}
-                                        getOptionLabel={(option) => option.streetAndNumber ?? ''}
-                                        error={formError}
-                                        isOptionEqualToValue={(option, value) => option?.bpna === value.bpna}
-                                        onChange={(_, value) =>
-                                            setTemporaryStock({ ...temporaryStock, stockLocationBpna: value?.bpna ?? undefined })
-                                        }
-                                        value={sites?.find(site => 
-                                            site.bpns === temporaryStock.stockLocationBpns)?.addresses?.find((a) => 
-                                                a.bpna === temporaryStock.stockLocationBpna) ?? null
-                                        }
-                                        label="Stock Address*"
-                                        placeholder="Select an address"
-                                        disabled={!temporaryStock.stockLocationBpns}
-                                        data-testid="stock-address-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <FormLabel>Quantity*</FormLabel>
-                                    <Input
-                                        id="quantity"
-                                        type="number"
-                                        placeholder="Enter quantity"
-                                        value={temporaryStock.quantity ?? ''}
-                                        error={formError && (temporaryStock?.quantity == null || temporaryStock.quantity < 0)}
-                                        onChange={(e) =>
-                                            setTemporaryStock((curr) => ({
-                                                ...curr,
-                                                quantity: e.target.value ? parseFloat(e.target.value) : undefined,
-                                            }))
-                                        }
-                                        sx={{ marginTop: '.5rem' }}
-                                        data-testid="stock-quantity-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        id="uom"
-                                        value={
-                                            temporaryStock.measurementUnit
-                                                ? {
-                                                      key: temporaryStock.measurementUnit,
-                                                      value: getUnitOfMeasurement(temporaryStock.measurementUnit),
-                                                  }
-                                                : null
-                                        }
-                                        options={UNITS_OF_MEASUREMENT}
-                                        getOptionLabel={(option) => option?.value ?? ''}
-                                        label="UOM*"
-                                        placeholder="Select unit"
-                                        error={formError && !temporaryStock?.measurementUnit}
-                                        onChange={(_, value) => setTemporaryStock((curr) => ({ ...curr, measurementUnit: value?.key }))}
-                                        isOptionEqualToValue={(option, value) => option?.key === value?.key}
-                                        data-testid="stock-uom-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <FormLabel>Customer Order Number</FormLabel>
-                                    <Input
-                                        id="customer-order-number"
-                                        type="text"
-                                        error={formError && !isValidOrderReference(temporaryStock)}
-                                        value={temporaryStock?.customerOrderNumber ?? ''}
-                                        onChange={(event) =>
-                                            setTemporaryStock({ ...temporaryStock, customerOrderNumber: event.target.value })
-                                        }
-                                        data-testid="stock-customer-order-number-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <FormLabel>Customer Order Position</FormLabel>
-                                    <Input
-                                        id="customer-order-position-number"
-                                        type="text"
-                                        error={formError && !isValidOrderReference(temporaryStock)}
-                                        value={temporaryStock?.customerOrderPositionNumber ?? ''}
-                                        onChange={(event) =>
-                                            setTemporaryStock({
-                                                ...temporaryStock,
-                                                customerOrderPositionNumber: event.target.value,
-                                            })
-                                        }
-                                        data-testid="stock-customer-order-position-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <FormLabel>Supplier Order Number</FormLabel>
-                                    <Input
-                                        id="supplier-order-number"
-                                        type="text"
-                                        value={temporaryStock?.supplierOrderNumber ?? ''}
-                                        onChange={(event) =>
-                                            setTemporaryStock({
-                                                ...temporaryStock,
-                                                supplierOrderNumber: event.target.value ?? '',
-                                            })
-                                        }
-                                        data-testid="stock-supplier-order-number-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6} alignContent="end">
-                                    <Stack direction="row" alignItems="center">
-                                        <Checkbox
-                                            id="isBlocked"
-                                            checked={temporaryStock?.isBlocked ?? false}
-                                            onChange={(event) => setTemporaryStock({ ...temporaryStock, isBlocked: event.target.checked })}
-                                            data-testid="stock-is-blocked-field"
-                                        />
-                                        <InputLabel htmlFor="isBlocked"> is Blocked </InputLabel>
-                                    </Stack>
-                                </Grid>
-                            </>
-                        ) : (
-                            <Grid item xs={12}>
-                                <Table
-                                    title={`Current ${stockType} stock`}
-                                    getRowId={(row) => row.uuid}
-                                    columns={createStockColumns(sites?.find(s => stocks.some(stock => stock.stockLocationBpns === s.bpns)) ? handleDelete : undefined)}
-                                    rows={stocks}
-                                    hideFooter
-                                    density="standard"
-                                />
-                            </Grid>
-                        )}
-                    </Grid>
-                    <Box display="flex" gap="1rem" width="100%" justifyContent="end" marginTop="2rem">
-                        <Button variant="outlined" color="primary" sx={{ display: 'flex', gap: '.25rem' }} onClick={onClose}>
-                            <Close></Close> Close
+                    <Table
+                        title={`Current ${stockType} stock`}
+                        getRowId={(row) => row.uuid}
+                        columns={createStockColumns(handleDelete, handleEdit)}
+                        rows={stocks}
+                        density="standard"
+                        hideFooter
+                    />
+                    <Box display="flex" justifyContent="flex-end" marginTop="2rem">
+                        <Button variant="outlined" onClick={onClose}>
+                            <Close /> Close
                         </Button>
-                        {mode === 'create' && (
-                            <Button variant="contained" color="primary" sx={{ display: 'flex', gap: '.25rem' }} onClick={handleSaveClick} data-testid="save-stock-button">
-                                <Save></Save> Save
-                            </Button>
-                        )}
                     </Box>
                 </Stack>
             </Dialog>